minor code reorganization and cleanup
